あちこちから使いそうなメソッドをUtilにまとめた。
（animateがUtilってほどどこでも使いそうかは微妙なので後で場所を変えるかもしれません）
