+++ app/pixi_renderer.js
@@ -3,6 +3,7 @@ import 'pixi.js';
 // import MessageLayer from 'message_layer';
 import LilyText from './lily_text';
 import Letter from './letter';
+import Util from './util';
 
 class PixiRenderer {
   constructor(){
@@ -80,7 +81,7 @@ class PixiRenderer {
     const sprite = this.sprites[label];
     if( sprite ) {
       const startAlpha = sprite.alpha;
-      await animate(duration, (rate)=>{
+      await Util.animate(duration, (rate)=>{
         sprite.alpha = startAlpha + rate * (alpha - startAlpha);
       });
     }
@@ -95,7 +96,7 @@ class PixiRenderer {
       const endX = left || sprite.x;
       const endY = top || sprite.y;
 
-      await animate(duration, (rate)=>{
+      await Util.animate(duration, (rate)=>{
         sprite.x = startX + rate * (endX - startX);
         sprite.y = startY + rate * (endY - startY);
       });
@@ -108,33 +109,6 @@ class PixiRenderer {
     delete this.sprites[label];
   }
 
-}
-
-// requestAnimationFrame wrapper, callbackは開始時からのdurationをミリ秒で受け取り、終わるとPromiseがresolveされます
-async function animate(duration, callback){
-  "use strict";
-  return new Promise(resolve => {
-    if(!(duration > 0)){
-      callback(1.0);
-      resolve();
-      return;
-    }
-    const start = Date.now();
-    const f = ()=> {
-      const current = Date.now();
-      const delta = current - start;
-      if( delta > duration ){
-        callback(1.0);
-        resolve();
-        return;
-      }
-      callback(delta / duration);
-
-      window.requestAnimationFrame(f);
-    };
-    f(start);
-  }
-  );
 }
 
 export default PixiRenderer;
